Move get_ligo_ldap_dns() to ldap_data.py
