+++ src/webapp/ldap_data.py
@@ -1,4 +1,9 @@
+import logging
+from typing import List
+
 import ldap3
+
+log = logging.getLogger(__name__)
 
 
 def get_contact_cilogon_id_map(global_data):
@@ -127,3 +132,37 @@ def merge_yaml_data(yaml_data_main, yaml_data_secondary):
 
     return yd
 
+
+def get_ligo_ldap_dns(ldapurl: str, ldapuser: str, ldappass: str) -> List[str]:
+    """
+    Query the LIGO LDAP server for all grid DNs in the IGWN collab.
+
+    Returns a list of DNs.
+    """
+    results = []
+    base_branch = "ou={group},dc=ligo,dc=org"
+    base_query = "(&(isMemberOf=Communities:{community})(gridX509subject=*))"
+    queries = {'people': base_query.format(community="LSCVirgoLIGOGroupMembers"),
+               'robot': base_query.format(community="robot:OSGRobotCert")}
+
+    try:
+        server = ldap3.Server(ldapurl, connect_timeout=10)
+        conn = ldap3.Connection(server, user=ldapuser, password=ldappass, raise_exceptions=True, receive_timeout=10)
+        conn.bind()
+    except ldap3.core.exceptions.LDAPException:
+        log.exception("Failed to connect to the LIGO LDAP")
+        return results
+
+    for group in ('people', 'robot'):
+        try:
+            conn.search(base_branch.format(group=group),
+                        queries[group],
+                        search_scope='SUBTREE',
+                        attributes=['gridX509subject'])
+            results += [dn for e in conn.entries for dn in e.gridX509subject]
+        except ldap3.core.exceptions.LDAPException:
+            log.exception("Failed to query LIGO LDAP for %s DNs", group)
+
+    conn.unbind()
+
+    return results
